pi/requirements.txt
picamera
python-dotenv
opencv-python
requests
pyserial
transliterate